requirements.txt
aiohttp==3.7.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.0.2
numpy==1.19.4
opencv-contrib-python==4.4.0.46
Pillow==8.0.1
pylint==2.6.0
pytesseract==0.3.6
requests==2.25.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
udpy==2.0.0
urbandictionary==1.1
urllib3==1.26.2
wrapt==1.12.1
yarl==1.6.3
